backend/python_server/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
crayons==0.4.0
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.2
opencv-python==4.13.0.92
piexif==1.1.3
pillow==11.3.0
stegano==2.2.0
Werkzeug==3.1.6
